refactor: tabs cn 파일 리팩터링 및 이름 수정
