+++ src/app/mypage/_components/StatsCard.tsx
@@ -2,14 +2,14 @@ type StatsCardProps = {
   title: string;
   value: number;
   total?: number;
-}
-const StatsCard = ({ title, value, total }: StatsCardProps) => {
+};
+const StatsCard = ({ title, value, total = 999 }: StatsCardProps) => {
   return (
-    <div className="rounded-lg bg-gray-100 p-6 text-center">
-      <h3 className="text-lg mb-2 font-medium">{title}</h3>
-      <p className="text-xl">
-        {value}
-        {total && <span className="text-gray-500"> / {total}</span>}
+    <div className="container inline-flex flex-col items-center justify-center gap-1 rounded-2xl p-6 text-center shadow-[0px_0px_1px_0px_rgba(0,0,0,0.08)] shadow-[0px_0px_8px_0px_rgba(0,0,0,0.12)]">
+      <h3 className="mb-2 text-title2 font-bold">{title}</h3>
+      <p className="text-title2 text-black">
+        <span className="text-primary-default">{value}</span>
+        {total && <span> / {total}</span>}
       </p>
     </div>
   );
