+++ src/app/mypage/page.tsx
@@ -4,10 +4,7 @@ import SignoutButton from "./_components/SignoutButton";
 
 const PrivatePage = () => {
   return (
-    <div className="h-screen w-full justify-items-center">
-      <div className="container mx-auto max-w-7xl my-8">
-        <h1 className="justify-self-start text-heading font-bold">마이페이지</h1>
-      </div>
+    <div className="h-screen mx-auto max-w-[62.25rem] justify-items-center">
       <ProfileSection />
       <MypageMenu />
       <SignoutButton />
